refactor(AuthForms): Grouping related forms in single folder
