+++ src/components/AuthForms/RegisterForm/index.ts
